requirements.txt
flask
flask-wtf
flask-login
flask-sqlalchemy
python-docx
pdfminer.six
python-magic
scikit-learn
pandas
numpy
spacy
nltk
email-validator